requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==25.1.0
beautifulsoup4==4.13.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
discord-webhook==1.3.0
discord.py==2.4.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
loguru==0.7.3
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
outcome==1.3.0.post0
packaging==24.2
propcache==0.2.1
pycparser==2.22
Pygments==2.19.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
rich==13.9.4
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.1
websocket-client==1.8.0
win32_setctime==1.2.0
wsproto==1.2.0
yarl==1.18.3
